Moved tag out of label
