+++ packages/react-components/src/button/stories/button.stories.mdx
@@ -1,9 +1,10 @@
 import { AddIcon, SignoutIcon } from "@react-components/icons";
 import { Button, PureButton } from "@react-components/button";
 import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
-import { Label, Tag } from "@react-components/label";
+import { Label } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
+import { Tag } from "@react-components/tag";
 import CustomColorHover from "./custom-color-hover.mdx";
 import Imports from "./imports.mdx";
