+++ packages/react-components/src/button/tests/chromatic/button/button.chroma.jsx
@@ -1,5 +1,6 @@
 import { Button } from "@react-components/button";
-import { Label, Tag } from "@react-components/label";
+import { Label } from "@react-components/label";
+import { Tag } from "@react-components/tag";
 import { createChromaticSection, paramsBuilder, storiesOfBuilder } from "@utils";
 import { createSharedStories } from "./shared-stories";
 import { isNil } from "lodash";
