+++ packages/react-components/src/label/src/factories.js
@@ -1,5 +1,4 @@
 import { Label } from "./label";
-import { Tag } from "./tag";
 import { isNil } from "lodash";
 
 export function createLabelFromShorthand({ content, ...props }) {
@@ -10,6 +9,4 @@ export function createLabelFromShorthand({ content, ...props }) {
     return <Label {...props} />;
 }
 
-export function createTagFromShorthand(props) {
-    return <Tag {...props} />;
-}
+
