+++ packages/react-components/src/label/stories/label.stories.mdx
@@ -1,9 +1,10 @@
 import { Button } from "@react-components/button";
 import { CloseIcon, CommunicationIcon, MagnifierIcon } from "@react-components/icons";
 import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
-import { Label, PureLabel, Tag } from "@react-components/label";
+import { Label, PureLabel } from "@react-components/label";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
+import { Tag } from "@react-components/tag";
 import { logButtonClicked } from "./utils";
 import Imports from "./imports.mdx";
