+++ packages/react-components/src/tag/index.js
@@ -1,0 +1,1 @@
+export * from "./src";
