+++ packages/react-components/src/tag/src/factories.js
@@ -1,0 +1,5 @@
+import { Tag } from "./tag";
+
+export function createTagFromShorthand(props) {
+    return <Tag {...props} />;
+}
