+++ packages/react-components/src/tag/stories/tag.stories.mdx
@@ -1,6 +1,6 @@
 import { HocWarning, InstallReactComponents, Props, ReactComponentsLayout } from "@blocks";
 import { Meta, Preview, Story } from "@storybook/addon-docs/blocks";
-import { PureTag, Tag } from "@react-components/label";
+import { PureTag, Tag } from "@react-components/tag";
 import { RefLogger, createComponentSection, paramsBuilder } from "@utils";
 import Imports from "./imports.mdx";
